+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Features
+
+* added toplevel activate
+
 ## [0.1.0] - 2025-08-21
 
 Initial Release